requirements.txt
pytest
pandas
numpy
ipython
ipykernel
jupyter_contrib_nbextensions
tqdm
jupyter_dashboards
spacy
scikit-learn
pdfminer.six
